# specs/012-food-api/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Biotrackr Food API
  description: |
    RESTful API for querying food and nutrition data from the Biotrackr health tracking system.
    
    This API provides read-only access to daily food logs, including calorie tracking, 
    water intake, and nutrition goals. Data is collected by the Food Service from Fitbit 
    and stored in Azure Cosmos DB.
    
    ## Features
    - Query food logs by specific date
    - Query food logs by date range with pagination
    - Get all food logs with pagination
    - Health check endpoint for monitoring
    
    ## Authentication
    Requires Azure API Management subscription key (`Ocp-Apim-Subscription-Key` header).
  version: 1.0.0
  contact:
    name: Biotrackr Support
    url: https://github.com/willvelida/biotrackr
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: https://{apimGateway}/food
    description: Production environment via Azure API Management
    variables:
      apimGateway:
        default: api.biotrackr.com
        description: APIM gateway hostname
  - url: http://localhost:8080
    description: Local development environment

tags:
  - name: Food Logs
    description: Operations for querying food and nutrition data
  - name: Health
    description: Health check and monitoring endpoints

paths:
  /:
    get:
      tags:
        - Food Logs
      summary: Get all food logs
      description: |
        Returns a paginated list of all food logs ordered by date (descending).
        
        Supports pagination via `pageNumber` and `pageSize` query parameters.
        Default page size is 20, maximum is 100.
      operationId: getAllFoodLogs
      parameters:
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Successfully retrieved food logs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedFoodLogsResponse'
              examples:
                successWithData:
                  summary: Successful response with data
                  value:
                    data:
                      - id: 550e8400-e29b-41d4-a716-446655440000
                        date: "2025-11-11"
                        summary:
                          calories: 2500
                          caloriesIn: 2200
                          water: 2000
                        goals:
                          calories: 2400
                      - id: 650e8400-e29b-41d4-a716-446655440001
                        date: "2025-11-10"
                        summary:
                          calories: 2300
                          caloriesIn: 2100
                          water: 1800
                        goals:
                          calories: 2400
                    totalCount: 365
                    pageNumber: 1
                    pageSize: 20
                    totalPages: 19
                emptyResult:
                  summary: No food logs found
                  value:
                    data: []
                    totalCount: 0
                    pageNumber: 1
                    pageSize: 20
                    totalPages: 0
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  /{date}:
    get:
      tags:
        - Food Logs
      summary: Get food log by date
      description: |
        Returns the food log for a specific date.
        
        Date must be in YYYY-MM-DD format (e.g., 2025-11-11).
        Returns 200 OK with null data if no food log exists for the date.
      operationId: getFoodLogByDate
      parameters:
        - $ref: '#/components/parameters/Date'
      responses:
        '200':
          description: Successfully retrieved food log
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/FoodLogResponse'
                  - type: object
                    properties:
                      data:
                        type: 'null'
                    description: No food log found for the specified date
              examples:
                found:
                  summary: Food log found
                  value:
                    id: 550e8400-e29b-41d4-a716-446655440000
                    date: "2025-11-11"
                    summary:
                      calories: 2500
                      caloriesIn: 2200
                      water: 2000
                    goals:
                      calories: 2400
                notFound:
                  summary: No food log for date
                  value:
                    data: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  /range/{startDate}/{endDate}:
    get:
      tags:
        - Food Logs
      summary: Get food logs by date range
      description: |
        Returns food logs within the specified date range (inclusive) with pagination.
        
        Both dates must be in YYYY-MM-DD format.
        Start date must be before or equal to end date.
      operationId: getFoodLogsByDateRange
      parameters:
        - $ref: '#/components/parameters/StartDate'
        - $ref: '#/components/parameters/EndDate'
        - $ref: '#/components/parameters/PageNumber'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Successfully retrieved food logs in range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedFoodLogsResponse'
              examples:
                found:
                  summary: Food logs found in range
                  value:
                    data:
                      - id: 550e8400-e29b-41d4-a716-446655440000
                        date: "2025-11-11"
                        summary:
                          calories: 2500
                          caloriesIn: 2200
                          water: 2000
                        goals:
                          calories: 2400
                      - id: 650e8400-e29b-41d4-a716-446655440001
                        date: "2025-11-10"
                        summary:
                          calories: 2300
                          caloriesIn: 2100
                          water: 1800
                        goals:
                          calories: 2400
                    totalCount: 30
                    pageNumber: 1
                    pageSize: 20
                    totalPages: 2
                emptyRange:
                  summary: No food logs in range
                  value:
                    data: []
                    totalCount: 0
                    pageNumber: 1
                    pageSize: 20
                    totalPages: 0
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'

  /healthz/liveness:
    get:
      tags:
        - Health
      summary: Liveness health check
      description: |
        Returns 200 OK if the API is running and responsive.
        
        Used by Kubernetes/Container Apps for liveness probes.
        Does not check external dependencies (Cosmos DB).
      operationId: livenessCheck
      responses:
        '200':
          description: API is alive and responsive
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: Healthy
              example:
                status: Healthy

components:
  parameters:
    Date:
      name: date
      in: path
      required: true
      description: Date in YYYY-MM-DD format (e.g., 2025-11-11)
      schema:
        type: string
        format: date
        pattern: '^\d{4}-\d{2}-\d{2}$'
        example: "2025-11-11"
      
    StartDate:
      name: startDate
      in: path
      required: true
      description: Start date in YYYY-MM-DD format (inclusive)
      schema:
        type: string
        format: date
        pattern: '^\d{4}-\d{2}-\d{2}$'
        example: "2025-11-01"
      
    EndDate:
      name: endDate
      in: path
      required: true
      description: End date in YYYY-MM-DD format (inclusive)
      schema:
        type: string
        format: date
        pattern: '^\d{4}-\d{2}-\d{2}$'
        example: "2025-11-30"
      
    PageNumber:
      name: pageNumber
      in: query
      required: false
      description: Page number (1-indexed). Default is 1.
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
      
    PageSize:
      name: pageSize
      in: query
      required: false
      description: Number of items per page. Default is 20, maximum is 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
        example: 20

  schemas:
    FoodLogResponse:
      type: object
      description: Single food log with nutrition data and goals
      required:
        - id
        - date
        - summary
        - goals
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier (GUID)
          example: 550e8400-e29b-41d4-a716-446655440000
        date:
          type: string
          format: date
          description: Date of the food log (YYYY-MM-DD)
          example: "2025-11-11"
        summary:
          $ref: '#/components/schemas/FoodLogSummary'
        goals:
          $ref: '#/components/schemas/FoodLogGoals'

    FoodLogSummary:
      type: object
      description: Daily nutrition totals
      required:
        - calories
        - caloriesIn
        - water
      properties:
        calories:
          type: integer
          description: Total calories burned (activity + BMR)
          minimum: 0
          example: 2500
        caloriesIn:
          type: integer
          description: Calories consumed from food and beverages
          minimum: 0
          example: 2200
        water:
          type: integer
          description: Water intake in milliliters
          minimum: 0
          example: 2000

    FoodLogGoals:
      type: object
      description: Daily nutrition targets
      required:
        - calories
      properties:
        calories:
          type: integer
          description: Daily calorie intake goal
          minimum: 0
          example: 2400

    PaginatedFoodLogsResponse:
      type: object
      description: Paginated collection of food logs with metadata
      required:
        - data
        - totalCount
        - pageNumber
        - pageSize
        - totalPages
      properties:
        data:
          type: array
          description: Array of food log items for current page
          items:
            $ref: '#/components/schemas/FoodLogResponse'
        totalCount:
          type: integer
          description: Total number of food logs across all pages
          minimum: 0
          example: 365
        pageNumber:
          type: integer
          description: Current page number (1-indexed)
          minimum: 1
          example: 1
        pageSize:
          type: integer
          description: Number of items per page
          minimum: 1
          maximum: 100
          example: 20
        totalPages:
          type: integer
          description: Total number of pages (calculated)
          minimum: 0
          example: 19

    ErrorResponse:
      type: object
      description: Standard error response
      required:
        - error
        - statusCode
      properties:
        error:
          type: string
          description: Human-readable error message
          example: Invalid date format. Expected YYYY-MM-DD.
        statusCode:
          type: integer
          description: HTTP status code
          example: 400

  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidDate:
              summary: Invalid date format
              value:
                error: Invalid date format. Expected YYYY-MM-DD.
                statusCode: 400
            invalidRange:
              summary: Invalid date range
              value:
                error: Invalid date range. Start date must be before or equal to end date.
                statusCode: 400
            invalidPageSize:
              summary: Invalid page size
              value:
                error: Invalid pageSize. Must be between 1 and 100.
                statusCode: 400
            invalidPageNumber:
              summary: Invalid page number
              value:
                error: Invalid pageNumber. Must be greater than 0.
                statusCode: 400
    
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: An unexpected error occurred. Please try again later.
            statusCode: 500
    
    ServiceUnavailable:
      description: Database or dependent service unavailable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Database is temporarily unavailable. Please try again later.
            statusCode: 503

  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Ocp-Apim-Subscription-Key
      description: Azure API Management subscription key

security:
  - ApiKeyAuth: []